requirements.txt
Flask==3.1.0
Flask-Session==0.8.0
openai>=1.0.0
python-dotenv==1.0.1
pydantic>=2.0.0
gunicorn==23.0.0
Pillow>=10.0.0
pillow-heif>=0.13.0
